+++ src/resources/workers/observability/destinations.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 import { SinglePage } from '../../../pagination';
 
@@ -28,7 +29,7 @@ export class Destinations extends APIResource {
     params: DestinationCreateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<DestinationCreateResponse> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.post(`/accounts/${account_id}/workers/observability/destinations`, {
         body,
@@ -62,7 +63,7 @@ export class Destinations extends APIResource {
     params: DestinationUpdateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<DestinationUpdateResponse> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.patch(`/accounts/${account_id}/workers/observability/destinations/${slug}`, {
         body,
@@ -85,10 +86,20 @@ export class Destinations extends APIResource {
    * ```
    */
   list(
-    params: DestinationListParams,
+    params?: DestinationListParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<DestinationListResponsesSinglePage, DestinationListResponse>;
+  list(
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<DestinationListResponsesSinglePage, DestinationListResponse>;
+  list(
+    params: DestinationListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<DestinationListResponsesSinglePage, DestinationListResponse> {
-    const { account_id, ...query } = params;
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { account_id = this._client.accountId, ...query } = params;
     return this._client.getAPIList(
       `/accounts/${account_id}/workers/observability/destinations`,
       DestinationListResponsesSinglePage,
@@ -110,10 +121,19 @@ export class Destinations extends APIResource {
    */
   delete(
     slug: string,
-    params: DestinationDeleteParams,
+    params?: DestinationDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<DestinationDeleteResponse>;
+  delete(slug: string, options?: Core.RequestOptions): Core.APIPromise<DestinationDeleteResponse>;
+  delete(
+    slug: string,
+    params: DestinationDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<DestinationDeleteResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(slug, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.delete(
         `/accounts/${account_id}/workers/observability/destinations/${slug}`,
@@ -245,7 +265,7 @@ export interface DestinationCreateParams {
   /**
    * Path param: Your Cloudflare account ID.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param
@@ -284,7 +304,7 @@ export interface DestinationUpdateParams {
   /**
    * Path param: Your Cloudflare account ID.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param
@@ -311,7 +331,7 @@ export interface DestinationListParams {
   /**
    * Path param: Your Cloudflare account ID.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Query param
@@ -338,7 +358,7 @@ export interface DestinationDeleteParams {
   /**
    * Your Cloudflare account ID.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 Destinations.DestinationListResponsesSinglePage = DestinationListResponsesSinglePage;
